Separate out ReqHandle into its own header
