+++ Doxyfile
@@ -1,3 +1,3 @@
 PROJECT_NAME = eRPC
-INPUT = src/nexus.h src/rpc.h src/msg_buffer.h src/rpc_types.h src/rpc_constants.h
+INPUT = src/nexus.h src/rpc.h src/msg_buffer.h src/rpc_types.h src/rpc_constants.h src/req_handle.h
 DISTRIBUTE_GROUP_DOC = YES
