+++ src/sslot.h
@@ -122,21 +122,4 @@ class SSlot {
   size_t get_cur_req_num() const { return cur_req_num_; }
 };
 
-/// Handle object passed by eRPC to the application's request handler callbacks
-class ReqHandle : public SSlot {
- public:
-  /// Get this RPC's request message buffer. Valid until the request handler
-  /// enqueues a response.
-  inline const MsgBuffer *get_req_msgbuf() const {
-    return &server_info_.req_msgbuf_;
-  }
-
-  /// Get the RPC ID of the server-side RPC object that received this request
-  /// from the network. Valid until the request handler enqueues a response.
-  uint8_t get_server_rpc_id() const;
-
-  /// Get the server-side session number for this request. Valid until the
-  /// request handler enqueues a response.
-  uint16_t get_server_session_num() const;
-};
 }  // namespace erpc
